refactor: 로그인 페이지를 조회하는 컨트롤러 변경
