--- src/main/java/roomescape/controller/LoginController.java
@@ -1,13 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-}
