+++ src/main/java/roomescape/controller/UserController.java
@@ -10,4 +10,9 @@ public class UserController {
     public String reservation() {
         return "reservation";
     }
+
+    @GetMapping("/login")
+    public String login() {
+        return "login";
+    }
 }
