--- src/test/java/roomescape/controller/LoginControllerTest.java
@@ -1,21 +1,0 @@
-package roomescape.controller;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class LoginControllerTest {
-
-    @Test
-    @DisplayName("로그인 페이지를 요청한다.")
-    void readLoginPage() {
-        RestAssured.given().log().all()
-                .when().get("/login")
-                .then().log().all()
-                .statusCode(200);
-    }
-}
