+++ src/test/java/roomescape/controller/UserControllerTest.java
@@ -1,10 +1,13 @@
 package roomescape.controller;
 
 import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
+
+import java.util.Map;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
@@ -15,6 +18,15 @@ class UserControllerTest {
     void reservation() {
         RestAssured.given().log().all()
                 .when().get("/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("로그인 페이지를 요청한다.")
+    void readLoginPage() {
+        RestAssured.given().log().all()
+                .when().get("/login")
                 .then().log().all()
                 .statusCode(200);
     }
